Added header guards for functionTracker.h. Also placed the class WyckPosTrackingInfo in its own file and changed the first letter of it to be capitalized (like most classes).
